requirements.txt
aioredis==1.3.1
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
async-timeout==4.0.0
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
Brlapi==0.6.6
certifi==2018.1.18
cffi==1.15.0
channels==3.0.4
channels-redis==2.4.2
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==35.0.0
cupshelpers==1.0
daphne==3.0.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
hiredis==2.0.0
httplib2==0.9.2
hyperlink==21.0.0
idna==2.6
incremental==21.3.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
msgpack==0.6.2
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
pathlib==1.0.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.26.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==21.0.0
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2021.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==21.1.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.4.2
system-service==0.3
systemd-python==234
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
xkit==0.0.0
zope.interface==5.4.0
